pyproject.toml
[build-system]
requires = ["setuptools >= 75.0", "setuptools_scm[toml]", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telemetry_rx"
dynamic = ["version"]
description = "Escort vehicle telemetry reciever, with support for USART & TCP/IP adapters."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE.md" }
authors = [
  { name = "Vorobiov Mihail", email = "mihail.vorobiov@gmail.com" },
  { name = "Dersedan Denisa", email = "dersedand@gmail.com" },
]
keywords = ["telemetry", "solis"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
  "cantools~=40.2.0",
  "click~=8.1.0",
  "crc~=7.1.0",
  "influxdb-client[ciso]~=1.48.0",
  "pyserial==3.5",
]

[project.urls]
Documentation = "https://github.com/VorobiovM/telemetry_rx"

[project.optional-dependencies]
dev = ["pre-commit", "pytest", "ruff"]

[project.scripts]
tool = "telemetry_rx.__main__:main"

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"
fallback_version = "0.0.1.dev0+unknown"

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]
line-length = 130
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["--import-mode=importlib"]
pythonpath = ["telemetry_rx"]
testpaths = ["tests"]
